+++ mycodo/mycodo_flask/authentication_routes.py
@@ -18,7 +18,10 @@ from flask_babel import gettext
 from flask.blueprints import Blueprint
 
 # Classes
-from mycodo.databases.mycodo_db.models import Misc
+from mycodo.databases.mycodo_db.models import (
+    AlembicVersion,
+    Misc
+)
 from mycodo.databases.users_db.models import Users
 
 # Functions
@@ -246,8 +249,28 @@ def authenticate_cookies(db_path, users):
     return False
 
 
+def check_database_version_issue():
+    alembic_version = db_retrieve_table(
+        current_app.config['MYCODO_DB_PATH'], AlembicVersion, entry='all')
+    if len(alembic_version) > 1:
+        flash("A check of your database indicates there is an issue with your"
+              " database version number. This issue first appeared in early "
+              "4.1.x versions of Mycodo and has since been resolved. However,"
+              " even though things may seem okay, this issue prevents your "
+              "database from being upgraded properly. Therefore, if you "
+              "continue to use Mycodo without regenerating your database, you"
+              " will assuredly experience issues. To resolve this issue, move"
+              " your mycodo.db from ~/Mycodo/databases/mycodo.db to a "
+              "different location (or delete it) and a new database will be "
+              "generated in its place. You will need to configure Mycodo from"
+              " scratch, but this is the only way to ensure your database is "
+              "able to be upgraded when the time comes. Sorry for the "
+              "inconvenience.", "error")
+
+
 def logged_in():
     """Verify the user is logged in"""
+    check_database_version_issue()
     if (not session.get('logged_in') and
             not authenticate_cookies(
                 current_app.config['USER_DB_PATH'], Users)):